Simplify the local ecli runner API
