+++ ecli/client/src/native_client.rs
@@ -9,10 +9,7 @@ use std::time::Duration;
 
 use ecli_lib::{
     config::ProgramType,
-    runner::{
-        client::{native::EcliNativeClient, AbstractClient},
-        LogType,
-    },
+    runner::{native::NativeRunner, native::RunningProgram, LogType},
 };
 
 use crate::helper::load_prog_buf_and_guess_type;
@@ -25,35 +22,29 @@ pub(crate) async fn run_native(
     args: &[String],
     user_prog_type: Option<ProgramType>,
 ) -> anyhow::Result<()> {
-    let client = EcliNativeClient::default();
     let (buf, prog_type) = load_prog_buf_and_guess_type(&prog, user_prog_type).await?;
 
-    let handle = client
-        .start_program(
-            Some("NativeProgram".to_string()),
-            &buf,
-            prog_type,
-            export_json,
-            args,
-            None,
-        )
-        .await?;
+    let program = NativeRunner::start_program(&buf, prog_type, export_json, args, None)?;
     let mut last_poll = None;
     loop {
-        let logs = client.fetch_logs(handle, last_poll, None).await?;
-        for (cursor, log) in logs.into_iter() {
-            if let LogType::Plain = log.log_type {
-                println!("{}", log.log);
-            } else {
-                print!("{}", log.log);
-            }
-            last_poll = Some(cursor + 1);
-        }
-        if TERMINATED.load(Ordering::Relaxed) {
+        print_new_logs(&program, &mut last_poll);
+        if TERMINATED.load(Ordering::Relaxed) || program.has_exited() {
+            print_new_logs(&program, &mut last_poll);
             break;
         }
         tokio::time::sleep(Duration::from_millis(100)).await;
     }
-    client.terminate_program(handle).await?;
+    program.terminate()?;
     Ok(())
+}
+
+fn print_new_logs(program: &RunningProgram, last_poll: &mut Option<usize>) {
+    for (cursor, log) in program.fetch_logs(*last_poll, None) {
+        if let LogType::Plain = log.log_type {
+            println!("{}", log.log);
+        } else {
+            print!("{}", log.log);
+        }
+        *last_poll = Some(cursor + 1);
+    }
 }
